Decompose src/diagnostics/checks.test.ts below 500-line limit (warren-7a15)
Extract shared captureSpawnCalls helper into checks.test-helpers.ts and
split the 10 describe blocks into sibling files: checks.test.ts keeps
bwrap/clone/clean, checks.config.test.ts holds config/db checks, and
checks.preview.test.ts holds preview checks. No test bodies altered.

53 pass / 0 fail; all files < 500 lines.
